+++ game/src/communication/clock/EventInterfaces.ts
