+++ game/src/communication/race/AnswerDTO.ts
@@ -1,0 +1,5 @@
+export interface AnswerDTO {
+	id: number;
+	label: string;
+	isRight: boolean;
+}
