+++ game/src/communication/race/EventInterfaces.ts
@@ -1,6 +1,10 @@
 import { ItemType } from "../../gameCore/race/items/Item";
-import ItemState from "./ItemState";
-import PlayerState from "./PlayerState";
+import { AnswerDTO } from "./AnswerDTO";
+import PlayerState, { PlayerEndState } from "./PlayerState";
+import { QuestionDTO } from "./QuestionDTO";
+import { StartingRaceGridInfo } from "./StartingGridInfo";
+
+//This file contains interfaces to type events that happen between the client and the server for RaceGame.
 
 export interface ItemUsedEvent {
 	itemType: ItemType;
@@ -27,17 +31,12 @@ export interface GameCreatedEvent {
 	isSinglePlayer: boolean;
 }
 
-//this includes the games options selected by the host
-export interface GameOptions {
-	gameTime: number;
+export interface PlayerLeftEvent {
+	playerId: string;
 }
 
 export interface GameEndEvent {
 	playerEndStates: PlayerEndState[];
-}
-
-export interface PlayerLeftEvent {
-	playerId: string;
 }
 
 //Maybe rework targetlocation to put it somewhere else ?
@@ -56,52 +55,10 @@ export interface AnswerCorrectedEvent {
 	targetLocation: Point;
 }
 
-export interface HostChangeEvent {
-	newHostName: string;
-}
-
 export interface QuestionAnsweredEvent {
 	playerId: string;
 	clientTimestamp: number;
 	answerTimestamp: number;
 	targetLocation: Point;
 	answer: AnswerDTO;
-}
-
-export interface PlayerEndState {
-	playerId: string;
-	points: number;
-	name: string;
-	//TODO: Maybe put the user information here
-	//to know what character model is used and possibly the grade the player is
-}
-
-export interface StartingRaceGridInfo {
-	width: number;
-	height: number;
-	nonWalkablePositions: Point[];
-	startingPositions: Point[];
-	finishLinePositions: Point[];
-	itemStates: ItemState[];
-}
-
-export interface QuestionDTO {
-	id: number;
-	answers: AnswerDTO[];
-	answerType: string;
-	schoolGradeId: number;
-	difficulty: number;
-	questionRelativePath: string;
-	feedbackRelativePath: string;
-}
-
-export interface AnswerDTO {
-	id: number;
-	label: string;
-	isRight: boolean;
-}
-
-export interface InfoForQuestion {
-	schoolGrade: number;
-	language: string;
 }
