+++ game/src/communication/race/ItemState.ts
@@ -1,5 +1,6 @@
 import { ItemType } from "../../gameCore/race/items/Item";
 
+//State used to locate item on the board
 export default interface ItemState {
 	type: ItemType;
 	location: Point;
