+++ game/src/communication/race/StartingGridInfo.ts
@@ -1,0 +1,10 @@
+import ItemState from "./ItemState";
+
+export interface StartingRaceGridInfo {
+	width: number;
+	height: number;
+	nonWalkablePositions: Point[];
+	startingPositions: Point[];
+	finishLinePositions: Point[];
+	itemStates: ItemState[];
+}
